+++ BusinessLogic/Services/Interfaces/IScheduleService.cs
@@ -4,5 +4,5 @@ namespace BusinessLogic.Services.Interfaces;
 
 public interface IScheduleService
 {
-    Task<IEnumerable<LessonDto>> GetScheduleForDate(DateTimeOffset dateTime, int userTelegramId);
+    Task<IEnumerable<LessonDto>> GetScheduleForDate(DateTimeOffset dateTime, long userTelegramId);
 }
